add parameters to `From`

+++ tests/derives/from.rs
@@ -1,5 +1,5 @@
 /* Crate imports */
 use nnn::nnn;
 
-#[nnn(nnn_derive(From))]
-struct Float(f64);
+#[nnn(nnn_derive(From, From<f64>))]
+struct Float(f32);
